New model integrated for add query and get query.
